refactor(condition): Moved all operators to subfolder
