+++ src/filter/rule/condition/operators/ConditionEqual.ts
@@ -1,13 +1,13 @@
-import { Condition } from './Condition';
-import { ConditionResult, ConditionSatisfied } from './ConditionResult';
+import { Condition } from '../Condition';
+import { ConditionResult, ConditionSatisfied } from '../ConditionResult';
 
-export class ConditionNotEqual extends Condition {
+export class ConditionEqual extends Condition {
 	check(inputValue): ConditionResult {
 		// Convert the input
-		const value = ConditionNotEqual.convertValue(inputValue);
+		const value = ConditionEqual.convertValue(inputValue);
 
 		// Check condition
-		if (value !== this.expectedValue) {
+		if (value === this.expectedValue) {
 			return new ConditionResult(this, value, ConditionSatisfied.YES);
 		}
