+++ src/filter/rule/condition/operators/ConditionGreaterThanOrEqual.ts
@@ -1,5 +1,5 @@
-import { Condition } from './Condition';
-import { ConditionResult, ConditionSatisfied } from './ConditionResult';
+import { Condition } from '../Condition';
+import { ConditionResult, ConditionSatisfied } from '../ConditionResult';
 
 export class ConditionGreaterThanOrEqual extends Condition {
 	check(inputValue): ConditionResult {
