+++ src/filter/rule/condition/operators/ConditionLessThan.ts
@@ -1,5 +1,5 @@
-import { Condition } from './Condition';
-import { ConditionResult, ConditionSatisfied } from './ConditionResult';
+import { Condition } from '../Condition';
+import { ConditionResult, ConditionSatisfied } from '../ConditionResult';
 
 export class ConditionLessThan extends Condition {
 	check(inputValue): ConditionResult {
